Consolidate GTK mocks
This was split across three files, consolidate into one.
